patch: non leader config servers also update their IP allow list
